#6: build: rewrite find_package_local to have extra search paths
The primary problem was that when a <pkg-name>_DIR is set, cmake uses
it in find_package and actually *changes* the value of the variable to
the directory actually found!

Thus,

  find_package_local(package "${package_DIR}/cmake/package" ...)

works properly the first time, but then fails when cmake is
re-run. This bug was very confusing because cmake changed the
behaviour with <pkg-name>_DIR being set.

Thus, find_package_local now takes alterative paths and automatically
tries multiple until one is found that works:

  find_package_local(
    package "${package_DIR}" Package
    "cmake/package" "CMake/package" # New arguments
  )