#631: trace: fix two-layed trace event causal connections
With a single event in the envelope, when tracing of system events is
enabled, the collection component used the envelope to connect the
true semantic "logical" connection for the event that caused a
broadcast (for instance). Add a new trace event to collection message
that holds the semantic connection so both can be preserved when
system events are being traced.

Now, if system tracing is enabled, the user will get the literal
connections from the broadcast (bouncing off of the root).