#723: term: add optional labels to epochs
Add a new optional parameter when creating a new epoch to specify a
label that is dumped in debugging output (DOT files). In order to make
this change correctly, a breaking change is made on non-default
overloads of TermminationDetector::makeEpochCollective and
TermminationDetector::makeEpochRooted due to char* to bool overload
problems.

Thus, a tag class is used to specify if the epoch should use the
current epoch as the succesor or whether a DS epoch should be
used. This has the added benefit of making the interface more explicit
and less error prone.