#665: message: revert serialized messenger to use base message (fix bug)
A bug was introduced when this call was converted to a MsgPtr due to
how the references are manipulated, cauing bad delete/frees.

With makeMessageImpl in the detail namespace we can call it for
special cases with hitting the deprecation warning for internal calls
that require a bare pointer for proper behaviour.