#882: component: implement general diagnostics for components
Composed in below all components is Diagnostic, which allows each
component to register a diagnostic with a key, description, etc. for
some type T.

At an exigent exit or after finalization, the diagnostic can be
traversed and then reduced across nodes to provide lots of useful
statistics for qualities registered for each component.