#912: CI: write new docker-compose targets for openmpi
Ideally, these wouldn't have to be specialty targets, instead the user
would just set an enviornment variable. However, in order to implement
that, all the current containers and image names would need to embed
the mpi implementation.

Thus, this is the easist way to support openmpi without changing the
entire configuration.