#1021: util: implement MaybeSerialize wrapper
The new wrapper class takes a fixed size and depending on the size of
T decides whether to serialize it or store it inline.

I still need to implement a new base type, MaybeSerializeBase<k>,
which will elide the type so that classes like CollectionMessage do
not need ColT or IndexT. Once that is done, we can register the
collection/index type so we can pick it up on the receive side and
dispatch to the collection.