#1134: cmake: import new GoogleTest*cmake w/3.18 and **fix bugs**
There is a major bug in the cmake googletest filter generation for
type-parameterized tests! This is causing some of our tests to never
run...including the ones that I just wrote which should always fail
without the associated fix.

Fix the regular expressions to generate a proper ctest filter that
will actually run *every* test