#1204: docs: use latest version and remove assertion from m.css
m.css has an assertion that limits the number of serialized children
and results used to build search strings by organizing them in a trie.
Commit 3fb8439 in VT introduced a new `serialize` method for
`LBManager`, which caused the number of children to exceed the amount
of storage allocated (4 bits or 16 children) due to the number of
related `serialize` methods. This commit updates VT to use the latest
version of m.css and removes the assertion during trie building for
search strings. This might cause strange behavior in searching, but
won't affect the rest of the documentation generation. A bug has been
submitted so we can hopefully remove this once they fix the bug:

https://github.com/mosra/m.css/issues/190