#1856: CI: This adds a script to run the build and test using cmake
I have not tested this as the compiler is failing
on a different .o file each run. I will set up the
changes needed in the vt-ci as well, but I don't
expect I will get it tested before leaving for vacation.
Note that use of a ctest script might  have been
a better pattern, but both are claimed to work in
the documentation.