#1856: Move to ctest script
Despite the documentation, the inclusion of ctest
calls in the cmake script was doing the build and
test, but not submitting anything to CDash. This moves
to a ctest script called using the -S flag. I am still
seeing test failures due to overloading, but it seems to
be caused by the test definitions itself. All the example
scripts use the macro add_test_for_example_vt which does
set the PROCESSORS property, but other tests use the add_test
macro directly and do not set the number of processors.
I found this by looking at the generated CTestTestfile.cmake
files.